feat: Implemented the commit log panel.
This implements a commit log panel that can be opened from both
diffviews and file history views. This will show the full commit
descriptions of all targeted commits. It can be configured to open
either as a floating window (default) or a normal split.

This commit also encompasses some major refactoring and extensions of
the 'Panel' class, which is the reason for the extensive list of changed
files.

Other changes:
- fix(file-history): The file history panel finally learned when not to
  use a plural 's'.
- fix(api): Fixed the order of execution of hooks in custom diffviews.