feat!: `:DiffviewFileHistory` with no args now targets the top-level.
Calling `:DiffviewFileHistory` with no args would previously target the
file in the current buffer. This has now been changed to instead target
the top-level of the working tree. This was changed because with how it
worked before, there was effectively no way to get the file history
equivalent of running `git log` with no path args. If your cwd was some
subdirectory of the working tree, and you wanted the full file history
of the tree, you would have to manually type out the path to the
top-level.

To get the file history for the file in the current buffer like before,
simply run: `:DiffviewFileHistory %`.

Other changes:
- You can now pass all the log options directly to the
  `DiffviewFileHistory` command, to override the default log options
  defined in the config.
- docs: More detailed documentation for file history and log options.
- fix: Fixed bug that made the commit log panel only work for a single
  working tree.
- feat(file-history): The git stats in the file panel are now aligned
  more nicely when there are commits with large change numbers.