feat!: More file history options (#151)
* feat(file-history): Added more flag options. Changed default log options

This makes all the options passed to the git-log command configurable.
You can now also provide the flag options directly to the
`:DiffviewFileHistory` command.

New flag options:
- `--first-parent`
- `--show-pulls`
- `--diff-merges`

* feat!: `:DiffviewFileHistory` with no args now targets the top-level.

Calling `:DiffviewFileHistory` with no args would previously target the
file in the current buffer. This has now been changed to instead target
the top-level of the working tree. This was changed because with how it
worked before, there was effectively no way to get the file history
equivalent of running `git log` with no path args. If your cwd was some
subdirectory of the working tree, and you wanted the full file history
of the tree, you would have to manually type out the path to the
top-level. On the contrary, getting the history for the current file is
always as simple as just using %, which expands to the current file
name.

To get the file history for the current file like before, simply run:
`:DiffviewFileHistory %`.

Other changes:
- You can now pass all the log options directly to the
  `DiffviewFileHistory` command, to override the default log options
  defined in the config.
- docs: More detailed documentation for file history and log options.
- fix: Fixed bug that made the commit log panel only work for a single
  working tree.
- feat(file-history): The git stats in the file panel are now aligned
  more nicely when there are commits with large change numbers.

* feat(file-history): Added `--reflog` and `++rev-range` options.

It's now possible to change the target rev range from within a file
history view as well. The option is called `++rev-range` in the option
panel, and it supports rev completion.

Other changes:
- fix: Ensure libuv timer handles are released.
- chore: Better error messages when the file history fails to update.
- chore: Adjusted some highlights.

* fix(file-history): Handle empty commit messages.

(ref. #150)

* docs: Amend changelog. Added deprecation warning.

* chore: Output coroutine errors in the log.