refactor!: Flexible layout system. (#190)
This is a huge refactor that gives the layout system in the views far
greater flexibility. The primary motivation behind this PR is the
planned merge-tool (#22) that is going to make use of 3-way diffs. I
want the conflicted files to appear in their own section in the normal
Diffview, thus the views need to be able to restructure the layout
dynamically depending on what entry is currently open. The previous
implementation was deeply tied to the 2-way diff layout, and it would be
unfeasible to force it to conform to a bodged dynamic layout system
without larger structural changes.

BREAKING CHANGES:

The only breaking change affects the action `view_windo`. This action is
not in use by default, so you won't be affected unless you know that
you've made use of it in your config.

Different entries in the same view may now use completely different
window layouts. Thus the action `view_windo` has changed to reflect
these changes. See `:h diffview-actions-view_windo` for more details on
the new usage.